+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 4.0.2 (UNRELEASED)
 * Added Italian translations.
+* Mask visibility=1 fields with asterisks in GetChanges json_data for unauthorized users.
 
 ## 4.0.1 (November 5, 2025)
 * Fixed longitude/latitude fields to be read-only when auto-geocoding is enabled